move get_item_marketplaces to tools.py
