+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -22,10 +22,11 @@ from demisto_sdk.commands.common.tools import (
     find_type, get_code_lang, get_current_repo, get_dict_from_file,
     get_entity_id_by_entity_type, get_entity_name_by_entity_type,
     get_file_displayed_name, get_file_version_suffix_if_exists,
-    get_files_in_dir, get_ignore_pack_skipped_tests, get_last_release_version,
-    get_last_remote_release_version, get_latest_release_notes_text,
-    get_pack_metadata, get_relative_path_from_packs_dir,
-    get_release_note_entries, get_release_notes_file_path, get_ryaml,
+    get_files_in_dir, get_ignore_pack_skipped_tests, get_item_marketplaces,
+    get_last_release_version, get_last_remote_release_version,
+    get_latest_release_notes_text, get_pack_metadata,
+    get_relative_path_from_packs_dir, get_release_note_entries,
+    get_release_notes_file_path, get_ryaml,
     get_scripts_and_commands_from_yml_data, get_test_playbook_id,
     get_to_version, get_yaml, has_remote_configured, is_origin_content_repo,
     is_pack_path, is_uuid, retrieve_file_ending, run_command_os,
@@ -1579,3 +1580,75 @@ def test_get_scripts_and_commands_from_yml_data(data, file_type, expected_comman
     commands, scripts = get_scripts_and_commands_from_yml_data(data=data, file_type=file_type)
     assert commands == expected_commands
     assert scripts == expected_scripts
+
+
+class TestGetItemMarketplaces:
+    @staticmethod
+    def test_item_has_marketplaces_field():
+        """
+        Given
+            - item declares marketplaces
+        When
+            - getting the marketplaces of an item
+        Then
+            - return the item's marketplaces
+        """
+        item_data = {
+            'name': 'Integration',
+            'marketplaces': ['xsoar', 'marketplacev2'],
+        }
+        marketplaces = get_item_marketplaces('Packs/PackID/Integrations/Integration/Integration.yml', item_data=item_data)
+
+        assert 'xsoar' in marketplaces
+        assert 'marketplacev2' in marketplaces
+
+    @staticmethod
+    def test_only_pack_has_marketplaces():
+        """
+        Given
+            - item does not declare marketplaces
+            - pack declares marketplaces
+        When
+            - getting the marketplaces of an item
+        Then
+            - return the pack's marketplaces
+        """
+        item_data = {
+            'name': 'Integration',
+            'pack': 'PackID',
+        }
+        packs = {
+            'PackID': {
+                'id': 'PackID',
+                'marketplaces': ['xsoar', 'marketplacev2'],
+            }
+        }
+        marketplaces = get_item_marketplaces('Packs/PackID/Integrations/Integration/Integration.yml', item_data=item_data, packs=packs)
+
+        assert 'xsoar' in marketplaces
+        assert 'marketplacev2' in marketplaces
+
+    @staticmethod
+    def test_no_marketplaces_specified():
+        """
+        Given
+            - item does not declare marketplaces
+            - pack does not declare marketplaces
+        When
+            - getting the marketplaces of an item
+        Then
+            - return the default marketplaces (only xsoar)
+        """
+        item_data = {
+            'name': 'Integration',
+            'pack': 'PackID',
+        }
+        packs = {
+            'PackID': {
+                'id': 'PackID',
+            }
+        }
+        marketplaces = get_item_marketplaces('Packs/PackID/Integrations/Integration/Integration.yml', item_data=item_data, packs=packs)
+
+        assert len(marketplaces) == 1
+        assert 'xsoar' in marketplaces
